fix additionnal line and variable name
